feat(deps): Update package name to recruitmentHRD and add @radix-ui/react-radio-group dependency (#12)
* feat(deps): Update package name to recruitmentHRD and add @radix-ui/react-radio-group dependency

* refactor(questions): Remove unused useEffect and clean up questions initialization

* refactor(questions): Simplify Questions component by removing props

* refactor: Remove unused imports and commented-out code in various components
