+++ resources/js/pages/candidate/settings/password.tsx
@@ -1,5 +1,4 @@
 import InputError from '@/components/input-error';
-import AppLayout from '@/layouts/app-layout';
 import SettingsLayout from '@/layouts/settings/user-layout';
 import { type BreadcrumbItem } from '@/types';
 import { Transition } from '@headlessui/react';
